Rename ircListeners.ts to ircClient.ts
