+++ lib/ircClient.ts
